+++ src/services/send_error_webhook.js
@@ -1,5 +1,5 @@
 import { get_webhook_data } from '../db_services/webhook_alert_dbservice.js'
-import {sendResponse} from './utils/alertUtils.js'
+import {sendResponse} from './utils/utilityService.js'
 
 async function send_error_to_webhook(bridge_id, org_id, error_log, error_type) {
   try {
